Refactor SignedVerdictOverviewForm into data and layout components
